+++ core/auth/auth.go
@@ -771,7 +771,7 @@ func (a *Auth) checkCredentials(identifierImpl identifierType, accountAuthType m
 	}
 
 	//if sign in was completed successfully, set auth type to verified
-	if accountAuthType.Unverified {	// && message != ""
+	if accountAuthType.Unverified {
 		accountAuthType.SetUnverified(false)
 		err := a.storage.UpdateAccountAuthType(accountAuthType)
 		if err != nil {
@@ -1666,9 +1666,6 @@ func (a *Auth) linkAccountAuthType(account model.Account, supportedAuthType mode
 			if err != nil {
 				return "", nil, err
 			}
-			// if message != "" {
-			// 	return "", nil, errors.ErrorData("incomplete", "verification", nil).SetStatus(utils.ErrorStatusUnverified)
-			// }
 			if aat != nil {
 				for i, accAuthType := range account.AuthTypes {
 					if accAuthType.ID == aat.ID {